Added test file filtering.
This commit adds the test file filtering system initially developed for
SoMoX4KAMP. It’s used to exclude test files from SoMoX’ analysis, in
order to obtain a more valuable results.

The commit includes some changes in the way filtering is handled in
order to easily plug in the filter. It contains no substantial changes
to the data and control flow.

git-svn-id: https://svnserver.informatik.kit.edu/i43/svn/code/SoMoX/Core/branches/SoMoXWithJaMoPP@33712 769cc9a0-fb0c-0410-8c2f-a2305f0e6268